bias/requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0




